junit test cases for #37
Since we are using functional reactive we need to declare web-application as reactive and similarly we need to set properties in @SpringBootTest

+++ mongodb-redis-integration/src/main/java/com/poc/mongodbredisintegration/document/Book.java
@@ -32,13 +32,10 @@ import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 /**
- * <p>
- * Book class.
- * </p>
+ * <p>Book class.</p>
  *
  * @author Raja Kolli
- * @version 0 : 5
- * @since July 2017
+ * @since 0.0.5
  */
 @Data
 @Document
@@ -62,6 +59,6 @@ public class Book implements Serializable {
 	private String text;
 
 	@Version
-	private int version;
+	private Long version;
 
 }
